+++ src/services/Fields.php
@@ -657,11 +657,13 @@ class Fields extends Component
     /**
      * Returns all fields that have a column in the content table.
      *
+     * @param string|string[]|false|null $context The field context(s) to fetch fields from. Defaults to [[\craft\services\Content::$fieldContext]].
+     * Set to `false` to get all fields regardless of context.
      * @return FieldInterface[] The fields
      */
-    public function getFieldsWithContent(): array
+    public function getFieldsWithContent($context = null): array
     {
-        return ArrayHelper::where($this->getAllFields(), function(FieldInterface $field) {
+        return ArrayHelper::where($this->getAllFields($context), function(FieldInterface $field) {
             return $field::hasContentColumn();
         }, true, true, false);
     }
@@ -669,12 +671,14 @@ class Fields extends Component
     /**
      * Returns all fields that don’t have a column in the content table.
      *
+     * @param string|string[]|false|null $context The field context(s) to fetch fields from. Defaults to [[\craft\services\Content::$fieldContext]].
+     * Set to `false` to get all fields regardless of context.
      * @return FieldInterface[] The fields
      * @since 3.7.60
      */
-    public function getFieldsWithoutContent(): array
+    public function getFieldsWithoutContent($context = null): array
     {
-        return ArrayHelper::where($this->getAllFields(), function(FieldInterface $field) {
+        return ArrayHelper::where($this->getAllFields($context), function(FieldInterface $field) {
             return !$field::hasContentColumn();
         }, true, true, false);
     }