+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Fixed an error that occurred if an arrow function was passed to the `|sort` Twig filter. ([#12334](https://github.com/craftcms/cms/issues/12334))
+- Fixed a bug where nested fields set to numbers could be inadvertently changed when an entry draft was created.
+- `craft\services\Fields::getFieldsWithContent()` and `getFieldsWithoutContent()` now have `$context` arguments.
 
 ## 3.7.60 - 2022-11-16